% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bookdown.R
\name{build_book}
\alias{build_book}
\title{Build the package bookdown documentation}
\usage{
build_book(output_format = "bookdown::gitbook")
}
\arguments{
\item{output_format}{Output format (default: "bookdown::gitbook")}
}
\description{
Build the package bookdown documentation
}
