% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clpmBias.R
\name{clpm_bias_matrix}
\alias{clpm_bias_matrix}
\title{Population CLPM slope bias from a random-intercept DGP (general form)}
\usage{
clpm_bias_matrix(omega, psi, b)
}
\arguments{
\item{omega}{Trait covariance matrix (k x k).}

\item{psi}{Within-person covariance matrix (k x k).}

\item{b}{Matrix of true within-person dynamics (k x k).}
}
\value{
A k x k matrix of CLPM slope biases.
}
\description{
CLPM bias is due to the nonzero correlation between the endogenous regressors and the omitted traits.
This function calculates the bias matrix for a general k-variable CLPM,
given the trait covariance matrix \eqn{\boldsymbol\Omega},
the within-person covariance matrix \eqn{\boldsymbol\Psi},
and the true within-person dynamics \eqn{\mathbf B^{\text{true}}}.
}
\keyword{internal}
