% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorIndicator.R
\name{generate_riclpm_indicator_lists}
\alias{generate_riclpm_indicator_lists}
\title{generate_riclpm_indicator_lists}
\usage{
generate_riclpm_indicator_lists(
  waves,
  num_indicators,
  has_z = FALSE,
  x_prefix = "x",
  y_prefix = "y",
  z_prefix = "z",
  indicator_separator = "_"
)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{num_indicators}{The number of indicators per construct (X, Y, and optionally Z) at each wave.}

\item{has_z}{Logical, indicating whether the model includes a Z variable. Defaults to FALSE.}

\item{x_prefix}{The prefix for the X variable indicators. Defaults to "x".}

\item{y_prefix}{The prefix for the Y variable indicators. Defaults to "y".}

\item{z_prefix}{The prefix for the Z variable indicators. Defaults to "z".}

\item{indicator_separator}{The separator between the prefix, indicator number, and wave number. Defaults to "_".}
}
\value{
A list containing the `time_varying_x`, `time_varying_y`, and optionally `time_varying_z` lists.
}
\description{
Generates lists of indicator variable names for a Random Intercept Cross-Lagged Panel Model (RICLPM).
}
\examples{
indicator_lists_2waves_3indicators <- generate_riclpm_indicator_lists(waves = 2, num_indicators = 3)
print(indicator_lists_2waves_3indicators)

indicator_lists_3waves_2indicators_withZ <- generate_riclpm_indicator_lists(waves = 3, num_indicators = 2, has_z = TRUE)
print(indicator_lists_3waves_2indicators_withZ)

indicator_lists_2waves_2indicators_custom_prefix <- generate_riclpm_indicator_lists(waves = 2, num_indicators = 2, x_prefix = "varX", y_prefix = "varY")
print(indicator_lists_2waves_2indicators_custom_prefix)

}
