% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLChange.R
\name{latentChange}
\alias{latentChange}
\title{Generate model syntax for a latent change score model}
\usage{
latentChange(waves = 10, model_type = c("latent_change", "dual_change"))
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{model_type}{Specify whether estimating a latent change score model for one variable or two.}
}
\value{
character string containing the Lavaan model syntax.
}
\description{
Generate model syntax for a latent change score model
}
