% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSim.R
\name{monteCarlo_OLS}
\alias{monteCarlo_OLS}
\title{simpleSim}
\usage{
monteCarlo_OLS(
  trials = 10,
  waves = 5,
  variance_between_x = 0.5,
  variance_between_y = 0.5,
  stability_q = 0.25,
  stability_p = 0.25,
  cross_p = 0,
  cross_q = 0,
  variance_p = 1,
  variance_q = 1,
  sample_size = 2500,
  ...
)
}
\arguments{
\item{trials}{The number of trials for the Monte Carlo simulation.}

\item{waves}{The number of waves (time points) in the model.}

\item{model}{Specify whether an "ols" or "clpm" model.}

\item{data_generation}{Specify the data generation method: "clpm", "ri-clpm"}

\item{proportion_change_x}{Proportion of change in x.}

\item{proportion_change_y}{Proportion of change in y.}
}
\value{
A data frame containing the results of the simulation.
}
\description{
Monte Carlo simulation comparing OLS and CLPM estimators.
Allows the user to specify different data-generating conditions (CLPM or RI-CLPM),
and apply either an OLS or CLPM model to the data.
}
