% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo_CTSEM.R
\name{monteCarloCTSEM}
\alias{monteCarloCTSEM}
\title{monteCarloCTSEM}
\usage{
monteCarloCTSEM(
  trials = 10,
  waves = 5,
  variance_between_x = 0.5,
  variance_between_y = 0.5,
  stability_q = 0.25,
  stability_p = 0.25,
  cross_p = 0,
  cross_q = 0,
  variance_p = 1,
  variance_q = 1,
  sample_size = 2500,
  dgp = "riclpm",
  cores = 10,
  confounder_p = 0.3,
  confounder_q = 0.3,
  confounder_variance = 1,
  confounder_stability = 0.4,
  include_confounder = TRUE,
  ...
)
}
\arguments{
\item{trials}{The number of trials for the Monte Carlo simulation.}

\item{waves}{The number of waves (time points) in the model.}

\item{model}{Specify whether an "ols" or "clpm" model.}

\item{data_generation}{Specify the data generation method: "clpm", "ri-clpm"}

\item{proportion_change_x}{Proportion of change in x.}

\item{proportion_change_y}{Proportion of change in y.}
}
\value{
A data frame containing the results of the simulation.
}
\description{
A somewhat strange function. It's common to estimate the cross-lagged panel model with two OLS regression models.
allows the user to specify different Data Generating conditions, and apply the OLS model to the data.
The output is a data frame that includes the estimated coefficients across these trials
}
