% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo_LChange.R
\name{monteCarloLChange}
\alias{monteCarloLChange}
\title{monteCarloRICLPM}
\usage{
monteCarloLChange(
  trials = 10,
  waves = 5,
  variance.between.x = 0.5,
  variance.between.y = 0.5,
  stability.q = 0.25,
  stability.p = 0.25,
  cross.p = 0,
  cross.q = 0,
  variance.p = 1,
  variance.q = 1,
  sample_size = 2500,
  dgp = "riclpm",
  ...
)
}
\arguments{
\item{trials}{The number of trials for the Monte Carlo simulation.}

\item{waves}{The number of waves (time points) in the model.}

\item{model}{Specify whether an "ols" or "clpm" model.}

\item{data_generation}{Specify the data generation method: "clpm", "ri-clpm"}

\item{proportion.change_x}{Proportion of change in x.}

\item{proportion.change_y}{Proportion of change in y.}
}
\value{
A data frame containing the results of the simulation.
}
\description{
A somewhat strange function. It's common to estimate the cross-lagged panel model with two OLS regression models.
allows the user to specify different Data Generating conditions, and apply the OLS model to the data.
The output is a data frame that includes the estimated coefficients across these trials
}
