% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datLong.R
\name{reshape_long_sim_cr}
\alias{reshape_long_sim_cr}
\title{reshape_long_sim_cr}
\usage{
reshape_long_sim_cr(data = dat)
}
\arguments{
\item{data}{A wide-format data frame containing variables named 'x1', 'x2', ..., 'xn', and 'y1', 'y2', ..., 'yn'.}
}
\value{
A long-format data frame with columns 'id', 'wave', 'x', 'y', 'xlag', and 'ylag'.
}
\description{
Reshape Data from Wide to Long Format for Cross-Lagged Analysis

This function takes a wide-format data frame containing variables named 'x1', 'x2', ..., 'xn', and 'y1', 'y2', ..., 'yn', and reshapes it into a long format suitable for cross-lagged analysis. The resulting data frame includes columns for individual ID, wave, x, y, xlag (lagged x), and ylag (lagged y).
}
\examples{


dat = simulate_observed_clr(waves = 3)$data

head(dat)

}
