% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCLPMu.R
\name{simCLPM_timeInvariantU}
\alias{simCLPM_timeInvariantU}
\title{simCLPM_timeInvariantU}
\usage{
simCLPM_timeInvariantU(waves = 10, ...)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{...}{Additional arguments passed to `lavaan::simulateData`.}
}
\value{
See [simCLPMu()].
}
\description{
Deprecated. Use
  `simCLPMu(confounder_type = "time_invariant")` instead.
}
\keyword{internal}
