% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsimOLS.R
\name{simulate_observed_clr}
\alias{simulate_observed_clr}
\title{Simulate data from an observed cross-lagged regression (CLR) model}
\usage{
simulate_observed_clr(
  waves = 10,
  stability.x = 0.9,
  stability.y = 0.9,
  contemporaneous.x = 0.5,
  cross.x = 0.3,
  beta.z = 0.5,
  beta.u = 0.5,
  var.y = 1,
  var.x = 1,
  cov.xy = 0.5,
  cov.uz = 0,
  ...
)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{stability.x}{The stability parameter for the x variable (autoregressive effect).}

\item{stability.y}{The stability parameter for the y variable (autoregressive effect).}

\item{contemporaneous.x}{The effect of the x variable on the y variable at the same time point.}

\item{cross.x}{The cross-lagged effect of x on y at the next time point.}

\item{beta.z}{The effect of the instrument (z) on the x variable.}

\item{beta.u}{The effect of the unobserved variable (u) on the observed variables (x and y).}

\item{var.y}{The variance of the y variable.}

\item{var.x}{The variance of the x variable.}

\item{cov.xy}{The covariance between x and y within the same time point.}

\item{cov.uz}{The covariance between the unobserved variable (u) and the instrument (z).}

\item{...}{Additional arguments to pass to the `lavaan::simulateData` function.}

\item{cross.y}{The cross-lagged effect of y on x at the next time point.}
}
\value{
A list containing two elements:
   * `model`: The Lavaan model syntax used for data simulation.
   * `data`: The simulated data in a data frame format.
}
\description{
Create a simulated dataset of a cross-lagged model with a specified number of waves and structural parameters.
}
