% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCLPM.R
\name{estimateCLPM}
\alias{estimateCLPM}
\title{model_syntax_clpm}
\usage{
estimateCLPM(waves = 10)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}
}
\value{
A character string containing the Lavaan model syntax for RICLPM.
   * `model`: The Lavaan model syntax used for data simulation.
}
\description{
Generates model syntax for  Cross-Lagged Panel Model (CLPM)
}
\details{
This function generates the model syntax for a Cross-Lagged Panel Model (CLPM)
with a specified number of waves.
}
\examples{
# Not Run
# Generate model syntax for a RICLPM with 5 waves
#  model_syntax = estimateCLPM(waves = 5)
#  cat(model_syntax)
}
