% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRICLPM.R
\name{estimateCLPM}
\alias{estimateCLPM}
\title{model_syntax_clpm}
\usage{
estimateCLPM(waves = 10)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{...}{Additional arguments to pass to the `lavaan` function.}
}
\value{
A character string containing the Lavaan model synthex for CLPM.
   * `model`: The Lavaan model syntax used for data simulation.
}
\description{
Generates model syntax for CLPM
}
