% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCTSEM.R
\name{estimateCTSEM}
\alias{estimateCTSEM}
\title{estimateCTSEM}
\usage{
estimateCTSEM(
  data = dat_long,
  id = "id",
  manifest.var.names = c("x", "y"),
  latent.var.names = c("x", "y"),
  LAMBDA = diag(2),
  type = "stanct",
  time_points = 10,
  chains = chains,
  ...
)
}
\arguments{
\item{data}{Must be a long format data frame.}

\item{id}{id, which is used to estimate subject specific effects.}
}
\value{
A list; returns model and drift matrix
#'
}
\description{
Create a continuous time model and generate a one unit discrete time
prediction for auto and cross effects.
The function returns the auto and cross effects translated to a discrete time model,
and a change of t, dt = 1. These are the AR AND CL parameters in the CLPM.

#'
}
