% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHLM.R
\name{estimateHLM}
\alias{estimateHLM}
\title{estimateHLM}
\usage{
estimateHLM(
  data = dat_long,
  residual_correlation = TRUE,
  chains = 3,
  cores = 10,
  iterations = 3000,
  ...
)
}
\arguments{
\item{data}{Long data frame, following structure from datLong() names() -> x, y, xlag, ylag}

\item{chains}{Specify the number of sampling chains, default 3}

\item{cores}{Number of cores to use for the model, default 10}

\item{iterations}{Number of iterations for the model, default 3000}

\item{residual_correlations}{Logical, should residual correlations be estimated?}
}
\value{
brms object
}
\description{
Generates model for HLM model in brms
}
\details{
This function generates a model for a Hierarchical Linear Model (HLM) using the brms package.
The data must be in long form, and include the lagged variables xlag and ylag.
The model includes random intercepts and slopes for x and y.
}
\examples{
# Not Run
# simRICLPM(waves = 5,
#           variance.p = 0.5,
#           variance.q = 0.5,
#           stability.p = 0.20,
#           stability.q = 0.20,
#           cross.q = 0.30,
#           cross.p = 0.30,
#           cov.pq = 0.00,
#           variance.between.x = 0.5,
#          variance.between.y = 0.5,
#           cov.between = 0,
#           sample.nobs = 10e2)[[2]] \%>\%
#   mutate(id = seq(1:nrow(df[[2]]))) \%>\%
#   reshape_long_sim_cr() -> dat_long
#
# estimateHLM()
}
