% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCLPM.R
\name{estimateRICLPM}
\alias{estimateRICLPM}
\title{estimateRICLPM}
\usage{
estimateRICLPM(waves = 10)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}
}
\value{
A character string containing the Lavaan model syntax for RICLPM.
   * `model`: The Lavaan model syntax used for data simulation.
}
\description{
Generates model syntax for Random Intercept Cross-Lagged Panel Model (RICLPM)
}
\details{
This function generates the model syntax for a Random Intercept Cross-Lagged Panel Model (RICLPM)
with a specified number of waves. The model includes latent variables for each wave, stability paths,
and covariances between the latent variables.
}
\examples{
# Not Run
# Generate model syntax for a RICLPM with 5 waves
#  model_syntax = estimateRICLPM(waves = 5)
#  cat(model_syntax)
}
