% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esimateRICLPM3.R
\name{estimateRICLPM3}
\alias{estimateRICLPM3}
\title{estimateRICLPM}
\usage{
estimateRICLPM3(
  data,
  time_varying_x,
  time_varying_y,
  time_varying_z,
  time_invariant_vars = NULL,
  waves = 10
)
}
\arguments{
\item{data}{A data frame containing the user's data.}

\item{time_varying_x}{A vector of variable names for the X variables in the data frame. Ordered in time from earliest to latest.}

\item{time_varying_y}{A vector of variable names for the Y variables in the data frame. Ordered in time from earliest to latest.}

\item{time_varying_z}{A vector of variable names for the Z variables in the data frame. Ordered in time from earliest to latest.}

\item{time_invariant_vars}{A vector of variable names for the time-invariant variables.}

\item{waves}{The number of waves (time points) in the model.}
}
\value{
A character string containing the Lavaan model syntax for RI-CLPM.
   * `model`: The Lavaan model syntax used for data simulation.
}
\description{
Generates model syntax for Random Intercept Cross-Lagged Panel Model (RICLPM)
}
\details{
This function generates the model syntax for a Random Intercept Cross-Lagged Panel Model (RICLPM)
with a specified number of waves. The model includes latent variables for each wave, stability paths,
and covariances between the latent variables.
}
\examples{
# Not Run
# Generate model syntax for a RICLPM with 5 waves
#  model_syntax = estimateRICLPM(data = my_data, time_varying_x = c("x1", "x2", "x3", "x4", "x5"), time_varying_y = c("y1", "y2", "y3", "y4", "y5"), time_varying_z = c("z1", "z2", "z3", "z4", "z5"), waves = 5, time_invariant_vars = c("w1", "w2"))
#  cat(model_syntax)
}
