% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRICLPM_nolag.R
\name{estimateRICLPM_nolag}
\alias{estimateRICLPM_nolag}
\title{estimateRICLPM_nolag}
\usage{
estimateRICLPM_nolag(
  time_varying_x,
  time_varying_y,
  time_varying_z = NULL,
  time_invariant_vars = NULL,
  waves
)
}
\arguments{
\item{time_varying_x}{A vector of variable names for the X variables, ordered from earliest to latest.}

\item{time_varying_y}{A vector of variable names for the Y variables, ordered from earliest to latest.}

\item{time_varying_z}{An optional vector of variable names for an additional Z variable, ordered from earliest to latest.}

\item{time_invariant_vars}{A vector of variable names for time-invariant variables.}

\item{waves}{The number of waves (time points) in the model.}
}
\value{
A character string containing the Lavaan model syntax for RICLPM.
}
\description{
Generates model syntax for Random Intercept Cross-Lagged Panel Model (RICLPM), no AR paths.
}
\details{
This function generates the model syntax for a Random Intercept Cross-Lagged Panel Model (RICLPM)
with a specified number of waves. The model includes latent variables for each wave, stability paths,
and covariances between the latent variables.
}
