% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRICLPMl.R
\name{estimateRICLPMl}
\alias{estimateRICLPMl}
\title{estimateRICLPMl}
\usage{
estimateRICLPMl(
  time_varying_x = list(c("x1_1", "x1_2"), c("x2_1", "x2_2")),
  time_varying_y = list(c("y1_1", "y1_2"), c("y2_1", "y2_2")),
  time_varying_z = NULL,
  time_invariant_vars = NULL,
  waves
)
}
\arguments{
\item{time_varying_x}{A list of vectors, each containing variable names for the X indicators at each wave. This is clunky, but basically declare the}

\item{time_varying_y}{A list of vectors, each containing variable names for the Y indicators at each wave.}

\item{time_varying_z}{An optional list of vectors, each containing variable names for the Z indicators at each wave.}

\item{time_invariant_vars}{A vector of variable names for time-invariant variables.}

\item{waves}{The number of waves (time points) in the model.}
}
\value{
A character string containing the Lavaan model syntax for RICLPM with freed factor loadings after the first indicator.
}
\description{
Generates model syntax for Random Intercept Cross-Lagged Panel Model (RICLPM) with multiple indicators per construct, where the first indicator of each latent variable has a fixed factor loading of 1, and subsequent indicators have free factor loadings.
}
\details{
This function generates the model syntax for a Random Intercept Cross-Lagged Panel Model (RICLPM)
with a specified number of waves and multiple indicators for each construct. Importantly, for each
latent variable at each wave, the factor loading of the first indicator is fixed to 1 for identification
purposes, while the factor loadings of subsequent indicators are freely estimated.
}
\examples{
\dontrun{
cat(model_syntax)
}

}
