% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo_OLS.R
\name{monteCarlo_OLS}
\alias{monteCarlo_OLS}
\title{monteCarlo_OLS}
\usage{
monteCarlo_OLS(
  trials = 10,
  waves = 5,
  data_generation = c("ri-clpm", "ri-clpms"),
  variance.between.x = 0.5,
  variance.between.y = 0.5,
  within_person_stability_x = 0,
  within_person_stability_y = 0,
  cross_lag_x = 0,
  cross_lag_y = 0,
  variance_x = 1,
  variance_y = 1,
  sample_size = 2500,
  cov.between = 0.5,
  confound = 0.5,
  ...
)
}
\arguments{
\item{trials}{The number of trials for the Monte Carlo simulation.}

\item{waves}{The number of waves (time points) in the model.}

\item{data_generation}{Specify the data generation method: "clpm", "ri-clpm"}

\item{model}{Specify whether an "ols" or "clpm" model.}

\item{proportion.change_x}{Proportion of change in x.}

\item{proportion.change_y}{Proportion of change in y.}
}
\value{
A data frame containing the results of the simulation.
}
\description{
A somewhat strange function. It's common to estimate the cross-lagged panel model with two OLS regression models.
allows the user to specify different Data Generating conditions, and apply the OLS model to the data.
The output is a data frame that includes the estimated coefficients across these trials
}
