% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo_LChange.R
\name{monteCarloLChange}
\alias{monteCarloLChange}
\title{monteCarloLChange}
\usage{
monteCarloLChange(
  trials = 10,
  waves = 5,
  dgp = "lchange",
  model_type = "dual_change",
  stability_q = 0.25,
  stability_p = 0.25,
  variance_between_y = 0.5,
  variance_between_x = 0.5,
  cross_p = 0,
  cross_q = 0,
  variance_p = 1,
  variance_q = 1,
  sample_size = 2500,
  cov_pq = 0.1,
  confounder_p = 0.3,
  confounder_q = 0.3,
  confounder_variance = 1,
  confounder_stability = 0.4,
  include_confounder = TRUE,
  confounder_type = "time_variant",
  ar_x = -0.15,
  ar_y = -0.2,
  cl_x = -0.1,
  cl_y = -0.1,
  change_x = 0.08,
  change_y = 0.08,
  phi_x = 0.15,
  phi_y = 0.15,
  initial_var_x = 1,
  initial_var_y = 1,
  constant_change_var_x = 0.5,
  constant_change_var_y = 0.5,
  residual_variance_x = 0.5,
  residual_variance_y = 0.5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{trials}{The number of trials for the Monte Carlo simulation.}

\item{waves}{The number of waves (time points) in the model. Must specify three or more waves for identification.}

\item{dgp}{Specify the data generation method: "riclpm", "clpm", "clpmu", "lchange". Are data generated under one of these regimes?}

\item{model_type}{Type of latent change model: "latent_change" or "dual_change". Use dual change with multiple variables}

\item{stability_q}{Autoregressive effect for q (y) variable. These parameters are only relevant when specifying an RICLPM or CLPM. They are **not** part of the dual change model}

\item{stability_p}{Autoregressive effect for p (x) variable. These parameters are only relevant when specifying an RICLPM or CLPM. They are **not** part of the dual change model}

\item{variance_between_y}{Between-person variance for y variable. Ony relevant for the RICLPM.}

\item{variance_between_x}{Between-person variance for x variable. Only relevant for RICLPM.}

\item{cross_p}{Cross-lagged effect of y on x. Only relevant to the CLPM and RICLPM specification.}

\item{cross_q}{Cross-lagged effect of x on y. Only relevant to the CLPM and RICLPM specification.}

\item{variance_p}{Within-person variance for p variable. Only relevant to the RICLPM and CLPM specification.}

\item{variance_q}{Within-person variance for q variable. Only relevant ot the RICLPM and CLPM specification}

\item{sample_size}{Sample size for simulation}

\item{cov_pq}{Covariance between p and q (for CLPM/CLPMU)}

\item{confounder_p}{Effect of confounder on x variables (for clpmu)}

\item{confounder_q}{Effect of confounder on y variables (for clpmu)}

\item{confounder_variance}{Variance of the confounder (for clpmu)}

\item{confounder_stability}{Autoregressive effect of confounder (for clpmu)}

\item{include_confounder}{Whether to include confounder in clpmu model}

\item{confounder_type}{Character string specifying confounder type for clpmu dgp.
Must be one of "time_variant" (default) or "time_invariant".}

\item{ar_x}{Proportional effect for X in latent change DGP}

\item{ar_y}{Proportional effect for Y in latent change DGP}

\item{cl_x}{Coupling effect in latent change DGP}

\item{cl_y}{Coupling effect in latent change DGP}

\item{change_x}{Change-to-change effect in latent change DGP}

\item{change_y}{Change-to-change effect in latent change DGP}

\item{phi_x}{Change autoregression in latent change DGP}

\item{phi_y}{Change autoregression in latent change DGP}

\item{initial_var_x}{Initial variance for X in latent change DGP}

\item{initial_var_y}{Initial variance for Y in latent change DGP}

\item{constant_change_var_x}{Constant change variance for X in latent change DGP}

\item{constant_change_var_y}{Constant change variance for Y in latent change DGP}

\item{residual_variance_x}{Measurement error for X in latent change DGP}

\item{residual_variance_y}{Measurement error for Y in latent change DGP}

\item{verbose}{Whether to print progress and error messages.}
}
\value{
A data frame containing the results of the simulation.
}
\description{
Monte Carlo simulation for Latent Change Score Models
This function combines the estimation and simulation of latent change models over a fixed number of trials.
The output is a data frame that includes the estimated coefficients across these trials
}
