% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCLPMu.R
\name{simCLPMu}
\alias{simCLPMu}
\title{simCLPMu}
\usage{
simCLPMu(
  waves = 10,
  stability_p = 0.2,
  stability_q = 0.2,
  cross_p = 0.1,
  cross_q = 0.1,
  variance_p = 1,
  variance_q = 1,
  cov_pq = 0.1,
  confounder_p = 0.3,
  confounder_q = 0.3,
  confounder_variance = 1,
  confounder_stability = 0.4,
  ...
)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{stability_p}{The stability parameter for the x variable (autoregressive effect).}

\item{stability_q}{The stability parameter for the y variable (autoregressive effect).}

\item{cross_p}{The cross-lagged effect of y on x at the next time point.}

\item{cross_q}{The cross-lagged effect of x on y at the next time point.}

\item{variance_p}{The variance of the p latent variable.}

\item{variance_q}{The variance of the q latent variable.}

\item{cov_pq}{The covariance between x and y within the same time point.}

\item{confounder_p}{The effect of the confounder u on x variables.}

\item{confounder_q}{The effect of the confounder u on y variables.}

\item{confounder_variance}{The variance of the confounder u.}

\item{confounder_stability}{The stability parameter for the confounder (autoregressive effect).}

\item{...}{Additional arguments to pass to the `lavaan::simulateData` function.}
}
\value{
A list containing two elements:
   * `model`: The Lavaan model syntax used for data simulation.
   * `data`:  The simulated data in a data frame format.
}
\description{
Simulate data from a cross-lagged panel model (CLPM) with a time-variant confounder.
}
